Remove consts from file manager
