+++ application/src/Installation/Task/AddDefaultSettingsTask.php
@@ -1,7 +1,7 @@
 <?php
 namespace Omeka\Installation\Task;
 
-use Omeka\File\Manager as FileManager;
+use Omeka\Form\SettingForm;
 use Omeka\Installation\Installer;
 use Omeka\Module;
 use Omeka\Service\Paginator;
@@ -11,8 +11,8 @@ class AddDefaultSettingsTask implements TaskInterface
     protected $defaultSettings = [
         'version' => Module::VERSION,
         'pagination_per_page' => Paginator::PER_PAGE,
-        'media_type_whitelist' => FileManager::MEDIA_TYPE_WHITELIST,
-        'extension_whitelist' => FileManager::EXTENSION_WHITELIST,
+        'media_type_whitelist' => SettingForm::MEDIA_TYPE_WHITELIST,
+        'extension_whitelist' => SettingForm::EXTENSION_WHITELIST,
     ];
 
     public function perform(Installer $installer)
